:building_construction: [#56] Separate SOAP client and request
